requirements.txt
torch>=2.0
pyyaml>=6.0
tqdm>=4.65
numpy>=1.24
# Optional: only needed for BPE / larger-corpus experiments (configs/kaggle_*)
tiktoken>=0.5
datasets>=2.14
